CEF控件离屏渲染模式：设置页面光标的代码，统一放在平台相关的类中实现（备注：使用SDL3时，只有Linux平台的光标正常，Windows/MacOS的光标都不正确，可以用CefBrowser示例程序验证）。

+++ duilib/CEFControl/CefWindowUtils_Windows.cpp
@@ -148,7 +148,7 @@ void SetCefWindowCursor(CefWindowHandle cefWindow, CefCursorHandle cursor)
     if ((cefWindow == nullptr) || (cursor == nullptr)) {
         return;
     }
-    ::SetClassLongPtr((HWND)cefWindow, GCLP_HCURSOR, static_cast<LONG>(reinterpret_cast<LONG_PTR>(cursor)));
+    ::SetClassLongPtr((HWND)cefWindow, GCLP_HCURSOR, reinterpret_cast<LONG_PTR>(cursor));
     ::SetCursor(cursor);
 }
